+++ sdk/openai/azure-ai-openai-assistants/CHANGELOG.md
@@ -6,9 +6,29 @@
 
 ### Breaking Changes
 
+- The following class have been removed:
+  - `FileListResponse`
+  - `OpenAIPageableListOfAssistant`
+  - `OpenAIPageableListOfAssistantFile`
+  - `OpenAIPageableListOfMessageFile`
+  - `OpenAIPageableListOfRunStep`
+  - `OpenAIPageableListOfThreadMessage`
+  - `OpenAIPageableListOfThreadRun` 
+
+- We've introduced `PageableList<T>` these classes were used, except for `FileListResponse` where we simply return `List<OpenAIFile>`.
+- If you are using `listFilesWithResponse` and need to manually deserialize the `BinaryData` in the response, you can still use `PageableList<T>` like so:
+```java
+client.listFilesWithResponse(requestOptions)
+    .getValue()
+    .toObject(new TypeReference<PageableList<OpenAIFile>>() {})
+    .getData();
+```
+
 ### Bugs Fixed
 
 ### Other Changes
+
+- On `uploadFile` method in `AssistantsClient` and `AssistantsAsyncClient`, it is required to set the "filename" of the file, via `setFilename` method in `FileDetails` class. The double quote character `"`, the newline charactor `0x0A`, the return charactor `0x0D` in "filename" would be escaped by the client library.
 
 ## 1.0.0-beta.1 (2024-02-07)